backend/pyproject.toml
[tool.poetry]
name = "astro-cms-backend"
version = "0.1.0"
description = "Backend for Mini CMS"
authors = ["Tomas <tomas@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.128.3"
uvicorn = "^0.40.0"
motor = "^3.7.1"
pydantic-settings = "^2.12.0"
python-dotenv = "^1.2.1"
python-multipart = "^0.0.22"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
